Deprecate qiskit.tools
This commit deprecates the qiskit.tools module. Most of the
functionality in this module will be removed in the Qiskit 1.0.0
release. The sole exception is the qiskit.tools.parallel module and it's
associated public function qiskit.tools.parallel_map which has been
moved to qiskit.utils.
