+++ releasenotes/notes/deprecate-tools-392582e5d3c698eb.yaml
@@ -1,0 +1,39 @@
+---
+deprecations:
+  - |
+    The ``qiskit.tools.jupyter`` module has been deprecated and will be removed
+    in Qiskit 1.0.0. This module is deprecated because the functionality in
+    this module is tied to the legacy ``qiskit-ibmq-provider`` package which is
+    no longer supported and also only supported :class:`.BackendV1`. If you're
+    using this functionality currently, similar jupyter tools exist in the
+    `qiskit-ibm-provider <https://github.com/Qiskit/qiskit-ibm-provider>`__
+    package which can be used instead. The documentation for this module can
+    be found here:
+
+    https://docs.quantum.ibm.com/api/qiskit-ibm-provider/ibm_jupyter
+
+  - |
+    The ``qiskit.tools.monitor`` module has been deprecated and will be removed
+    in Qiskit 1.0.0. This module is deprecated because the functionality in
+    this module is tied to the legacy ``qiskit-ibmq-provider`` package which is
+    no longer supported and also only supported :class:`.BackendV1`.
+  - |
+    The ``qiskit.tools.visualization`` module has been deprecated and will be
+    removed in Qiskit 1.0.0. This module was a legacy redirect from the original
+    location of Qiskit's visualization module and was moved to
+    :mod:`qiskit.visualization` in Qiskit 0.8.0. If you're still using this
+    path you can just update your imports from ``qiskit.tools.visualization``
+    to :mod:`qiskit.visualization`.
+  - |
+    The ``qiskit.tools.events`` module and the :func:`.progressbar` utility
+    it exposed has been deprecated and will be removed in the Qiskit 1.0.0
+    release. This module's functionality was not widely used and better covered
+    by dedicated packages such as `tqdm <https://github.com/tqdm/tqdm>`__.
+  - |
+    The ``qiskit.tools`` module has been deprecated and will be removed in
+    Qiskit 1.0.0. Except as noted in the release notes above for specific
+    submodules (``qiskit.tools.jupyter``, ``qiskit.tools.monitor``, ``qiskit.tools.events`` and
+    ``qiskit.tools.visualization``) the functionality in this module have
+    been migrated to :mod:`qiskit.utils`. If you're using any functionality
+    in this module you can update your imports from ``qiskit.tools`` to
+    :mod:`qiskit.utils`.
